+++ History.md
@@ -3,6 +3,7 @@
  * Replaced: urllib2 with requests package
  * Replaced: oauth2 with requests_oauthlib package
  * Added: optional timeout parameter
+ * Removed: support for Python 2.4/2.5
 
 0.4.0 / 2013-07-23
 ==================